requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
mcp>=1.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.115.0
uvicorn>=0.30.0
duckduckgo-search>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pypdf>=5.0.0
